fix(evaluation): run real ONNX inference in pose baselines, require authorized media
The mediapipe_opencv benchmark candidate previously wired golden/fixed
tensors as its forward_fn, so its reported latency and coverage never
reflected real model execution. Add model_downloader.py to download,
verify (size + SHA-256 against the pinned manifest), and cache ONNX
weights under the git-ignored weights/ directory, and load them via
cv2.dnn.readNetFromONNX with genuine net.forward() calls. Golden
tensors remain reserved for the decode/preprocessing unit tests in
test_opencv_adapters.py.

Running real inference surfaced two genuine bugs invisible under
mocked tensors:
- person_detection_mediapipe_v1.json's size_bytes was 1199015, a factor
  of 10 short of the real 11990159-byte artifact (verified against the
  upstream Git LFS pointer and a working mirror).
- The detector's box/score tensor assignment assumed a fixed output
  order; OpenCV 5's net.forward() returns them in the opposite order
  for this model. Fixed by dispatching on tensor shape instead of
  position. _decode_pose is now defensive against unrecognized output
  layouts (some ONNX export variants emit heatmap-shaped landmark
  heads this decoder does not yet support) and returns no landmarks
  rather than crashing.

Require real authorized media by default: run_pose_baseline_benchmark
now resolves each clip's authorized local file via
ControlledMediaSourceAdapter and raises AuthorizedMediaUnavailableError
with the exact missing path when it cannot, rather than silently
generating synthetic frames. A new explicit allow_synthetic flag (CLI:
--synthetic) permits mechanics/latency-only benchmarking, restricted to
clips declaring consent_scope=synthetic-no-person, and labels every
result as synthetic (BenchmarkEnvironment.is_synthetic, surfaced in the
report and CLI/notebook output) so it can never be mistaken for a real
evaluation. No authorized recorded clips exist yet, so 02_pose_baselines.ipynb
currently runs in synthetic mode; its coverage assertions only apply in
real mode, since the real model honestly reports 0% coverage on
synthetic non-person imagery.

Add real load/forward evidence (test_real_onnx_inference.py) proving
the pinned weights download, verify, load via cv2.dnn.readNetFromONNX,
and execute — independent of authorized clip availability. Note: the
pose model's net.forward() has been observed to occasionally raise a
native OpenCV DNN error under this build's experimental graph engine
(~1 in 30+ runs); this is a native-library-level flake, not application
logic, and is disclosed rather than hidden or retried.

Fix 22 ruff findings (import ordering, redundant int(round()) casts,
a self-comparison NaN check, a blind except narrowed to the specific
exception types this CLI path can raise, dict.items()->values() where
keys are unused, an unused unpacked variable, an unsorted __all__, and
an unparenthesized string concatenation).

+++ src/kinetiq_v_vision/infrastructure/inference/preprocessing.py
@@ -6,6 +6,7 @@ Implements preprocessing pipelines specified by:
 """
 
 from dataclasses import dataclass
+
 import cv2
 import numpy as np
 
@@ -107,8 +108,8 @@ class PersonDetectionPreprocessor:
 
         # 2. Scale ratio preserving aspect ratio
         ratio = min(self.target_height / orig_h, self.target_width / orig_w)
-        new_w = int(round(orig_w * ratio))
-        new_h = int(round(orig_h * ratio))
+        new_w = round(orig_w * ratio)
+        new_h = round(orig_h * ratio)
 
         resized = cv2.resize(rgb, (new_w, new_h), interpolation=cv2.INTER_LINEAR)
